Refactor to use a single bucket (#3)
The first working implementation used buckets to represent directories. This was easy to implement and efficient while testing. Unfortunately, Distribution creates a lot of different directories. Pushing the python:3 image to a fresh registry creates about 50 buckets. The problem is that every bucket creation on a JetStream cluster results in a leader election, which can take up to half a second per bucket. That brings uploading the python:3 image to a crawl, taking over 40 seconds.

Refactoring to use a single bucket instead solves this particular issue. Going to write a design on how to optimize this further.