Implement store.Metadata on BoltDB (#41)
Initial implementation of the Metadata store on top of BoltDB!

The registry is now persistent! 🎉

There's some stuff to clean up, because store.Metadata is still lacking methods for managing repositories. As such, repositories are created on-demand (and ad-hoc) inside store.Metadata methods that write to the store. A follow-up PR will add those methods, allowing the RegistryService interface to manage repositories, where that functionality belongs. If we keep doing ad-hoc repository creation, that's where it will be easy to implement cleanly without it leaking into everything else.

Also includes an update to mockery, which I shouldn't have done. Oh well.